GN-3 Add namespaces and interfaces, create Application class and use it
