+++ src/LocalizationDataBuilder/Business/MasterProcessorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace LocalizationDataBuilder\Business;
+
+use LocalizationDataBuilder\Config\Config;
+
+interface MasterProcessorInterface
+{
+    /**
+     * @param \LocalizationDataBuilder\Config\Config $config
+     *
+     * @return array
+     */
+    public function processLocaleMaster(Config $config): array;
+}
