+++ src/LocalizationDataBuilder/Business/ReplacementDataProcessorInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace LocalizationDataBuilder\Business;
+
+use LocalizationDataBuilder\Config\Config;
+
+interface ReplacementDataProcessorInterface
+{
+    /**
+     * @param array $localeMaster
+     * @param \LocalizationDataBuilder\Config\Config $config
+     *
+     * @return array
+     */
+    public function composeReplacementDataForLocales(
+        array $localeMaster,
+        Config $config
+    ): array;
+}
